backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
Pillow>=10.0.0
pywin32
mss
ollama>=0.4.0
openai-whisper>=20240930
watchdog>=3.0.0
psutil>=5.9.0
numpy>=1.20.0,<2.0.0
